fix: improving App.tsx
