+++ frontend/src/App.tsx
@@ -9,8 +9,8 @@ import {
 import { Separator } from '@/components/ui/separator';
 import Login from './Login';
 import Home from './Home';
+import { useLocalStorage } from './hooks/useLocalStorage';
 import { Toaster } from 'sonner';
-import { useEffect, useState } from 'react';
 import type { LeagueData } from './components/types/league_data';
 import { ThemeProvider } from '@/components/themes/theme_provider';
 import { ModeToggle } from './components/themes/mode_toggle';
@@ -20,23 +20,10 @@ import { BrowserRouter as Router, Routes, Route, Navigate, useNavigate } from 'r
 function App() {
   return (
     <Router>
+      <Toaster position="top-center" />
       <AppContent />
     </Router>
   );
-}
-
-function useLocalStorage<T>(key: string, initialValue: T | null) {
-  const [state, setState] = useState<T | null>(() => {
-    const stored = localStorage.getItem(key);
-    return stored ? JSON.parse(stored) : initialValue;
-  });
-
-  useEffect(() => {
-    if (state) localStorage.setItem(key, JSON.stringify(state));
-    else localStorage.removeItem(key);
-  }, [key, state]);
-
-  return [state, setState] as const;
 }
 
 function AppContent() {
@@ -46,15 +33,9 @@ function AppContent() {
   return (
     <ThemeProvider defaultTheme="dark" storageKey="vite-ui-theme">
       <div>
-        <Toaster position="top-center" />
-        <div className="relative flex items-center px-4 py-2 w-full">
-          {/* Title, absolutely centered */}
-          <h1 className="absolute left-1/2 transform -translate-x-1/2 text-3xl font-bold">
-            Fantasy League History Visualizer
-          </h1>
-
-          {/* Icons, right-aligned */}
-          <div className="ml-auto flex items-center gap-4">
+        <div className="flex items-center justify-center relative px-4 py-2 w-full">
+          <h1 className="text-3xl font-bold mx-auto">Fantasy League History Visualizer</h1>
+          <div className="absolute right-4 flex items-center gap-4">
             <Dialog>
               <DialogTrigger asChild>
                 <Button variant="ghost" size="icon" aria-label="Info" className="cursor-pointer">
@@ -68,18 +49,8 @@ function AppContent() {
                 <p>Paragraph content -- change later</p>
               </DialogContent>
             </Dialog>
-            <Button
-              asChild
-              variant="ghost"
-              size="icon"
-              aria-label="GitHub"
-              className="cursor-pointer"
-            >
-              <a
-                href="https://github.com/amolrairikar/espn-fantasy-league-analytics"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
+            <Button asChild variant="ghost" size="icon" aria-label="GitHub" className="cursor-pointer">
+              <a href="https://github.com/amolrairikar/espn-fantasy-league-analytics" target="_blank" rel="noopener noreferrer">
                 <Link className="h-5 w-5" />
               </a>
             </Button>
